+++ src/plugins/migration-schema/migration-helpers.ts
@@ -91,7 +91,10 @@ export function migrateDocumentData(
         if (doc === null) {
             return PROMISE_RESOLVE_NULL;
         }
-        doc._meta = meta;
+
+        if (meta) {
+            doc._meta = meta;
+        }
         return doc;
     });
 }