+++ src/rx-document.ts
@@ -256,6 +256,9 @@ export const basePrototype = {
     putAttachment() {
         throw pluginMissing('attachments');
     },
+    putAttachmentBase64() {
+        throw pluginMissing('attachments');
+    },
     getAttachment() {
         throw pluginMissing('attachments');
     },